refactor(TestBed): extract injection installer to satisfy max-lines
